Move index manager mocks to the main conftest.py
