+++ apps/conftest.py
@@ -1,8 +1,10 @@
 import os
+from unittest.mock import Mock, patch
 
 import pytest
 from django.db import connections
 
+from apps.service_providers.llm_service.index_managers import LocalIndexManager, RemoteIndexManager
 from apps.teams.utils import unset_current_team
 from apps.utils.factories.experiment import ExperimentFactory
 from apps.utils.factories.team import TeamFactory, TeamWithUsersFactory
@@ -21,6 +23,25 @@ def team_with_users(db):
 @pytest.fixture()
 def experiment(team_with_users, db):
     return ExperimentFactory(team=team_with_users)
+
+
+@pytest.fixture()
+def index_manager_mock():
+    # TODO CS: Rename to remote_index_manager_mock
+    index_manager = Mock(spec=RemoteIndexManager)
+    with patch("apps.service_providers.models.LlmProvider.get_remote_index_manager") as get_remote_index_manager:
+        index_manager.client = Mock()
+        get_remote_index_manager.return_value = index_manager
+        yield index_manager
+
+
+@pytest.fixture()
+def local_index_manager_mock():
+    index_manager = Mock(spec=LocalIndexManager)
+    with patch("apps.service_providers.models.LlmProvider.get_local_index_manager") as get_local_index_manager:
+        index_manager.client = Mock()
+        get_local_index_manager.return_value = index_manager
+        yield index_manager
 
 
 @pytest.fixture(autouse=True, scope="session")
